re_view_graph::layout
Module slots

Force-directed graph layouts assume edges to be straight lines. A Slot represents the space that a single edge, or multiple edges can occupy between two nodes.

We achieve this by bringing edges into a canonical form via SlotId, which we can then use to find duplicates.

Structs

Slot

SlotId

Uniquely identifies a Slot by ordering the NodeIds that make up an edge.

Enums

SlotKind

There are different types of Slots that are laid out differently.

Functions

slotted_edges

Converts a list of edges into their slotted form.