re_view_graph
Module layout

Re-exports

pub use geometry::EdgeGeometry;

pub use geometry::PathGeometry;

pub use params::ForceLayoutParams;

pub use provider::ForceLayoutProvider;

pub use request::EdgeTemplate;

pub use request::LayoutRequest;

pub use result::Layout;

Modules

geometry 🔒

Provides geometric (shape) abstractions for the different elements of a graph layout.

params 🔒

provider 🔒

Performs the layout of the graph, i.e. converting an LayoutRequest into a Layout.

request 🔒

Contains all the (geometric) information that is considered when performing a graph layout.

result 🔒

Defines the output of a layout algorithm, i.e. everything that we need to render the graph.

slots 🔒

Force-directed graph layouts assume edges to be straight lines. A Slot represents the space that a single edge, or multiple edges can occupy between two nodes.